wip, service struct in config, vcap_services handling
