+++ config/const.go
@@ -1,5 +1,6 @@
 package config
 
+const BackmanConfig = "BACKMAN_CONFIG"
 const BackmanUsername = "BACKMAN_USERNAME"
 const BackmanPassword = "BACKMAN_PASSWORD"
 const BackmanEncryptionKey = "BACKMAN_ENCRYPTION_KEY"
